Fix issue w/ Cent/RHEL 6/7 causing sshd to fail startup
CentOS/RHEL 7 require AuthorizedKeysCommandUser if AuthorizedKeysCommand is used. 6 however will fail if this value is present. I am unsure what debian, various version and variants will do, so adding server.authorized_keys_command.user is advantageous in that you will be able to have the option of this being present or not even if you can't span osmajorversions.